+++ src/util.ts
@@ -511,52 +511,6 @@ class Util {
         disposables.splice(0, disposables.length);
     }
 
-    public async convertRegistryEcpResponseToScope(response: Response): Promise<EcpRegistryData> {
-        if (typeof response.body === 'string') {
-            try {
-                let parsed = await this.parseXml(response.body) as RegistryAsJson;
-
-                const status = parsed['plugin-registry'].status[0];
-
-                if (status === 'OK') {
-
-                    let registry = parsed['plugin-registry'].registry?.[0];
-
-                    let sections: EcpRegistryData['sections'] = {};
-                    for (const section of registry?.sections?.[0]?.section ?? []) {
-                        if (typeof section === 'string') {
-                            continue;
-                        }
-                        let sectionName = section.name[0];
-                        for (const item of section.items[0].item) {
-                            sections[sectionName] ??= {};
-                            sections[sectionName][item.key[0]] = item.value[0];
-                        }
-                    }
-
-                    return {
-                        devId: registry?.['dev-id']?.[0],
-                        plugins: registry?.plugins?.[0]?.split(','),
-                        sections: sections,
-                        spaceAvailable: registry?.['space-available']?.[0],
-                        status: 'OK'
-                    };
-                } else {
-                    return {
-                        status: 'FAILED',
-                        errorMessage: parsed?.['plugin-registry']?.error?.[0] ?? 'Unknown error'
-                    };
-                }
-            } catch {
-                //if the response is not xml, just return the body as-is
-                return {
-                    status: 'FAILED',
-                    errorMessage: response.body
-                };
-            }
-        }
-    }
-
     /**
      * Parse an xml file and get back a javascript object containing its results
      */
@@ -571,38 +525,6 @@ class Util {
             });
         });
     }
-}
-
-export interface EcpRegistryData {
-    devId?: string;
-    plugins?: Array<string>;
-    sections?: Record<string, Record<string, string>>;
-    spaceAvailable?: string;
-    status: 'OK' | 'FAILED';
-    errorMessage?: string;
-}
-
-interface RegistryAsJson {
-    'plugin-registry': {
-        registry: [{
-            'dev-id': [string];
-            plugins: [string];
-            sections: [{
-                section: [{
-                    items: [{
-                        item: [{
-                            key: [string];
-                            value: [string];
-                        }];
-                    }];
-                    name: [string];
-                } | string];
-            }];
-            'space-available': [string];
-        }];
-        status: [string];
-        error?: [string];
-    };
 }
 
 export function defer<T>() {
