+++ tsconfig.json
@@ -14,6 +14,7 @@
     "preserveConstEnums": true,
     "strictNullChecks": false,
     "noUnusedParameters": false,
+    "strictBindCallApply": true,
     "lib": [
         "es2017",
         "es2019.array",
